refactor: tests, format
