+++ src/helpers/format.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck
 import * as YAML from "yaml"
 import { getInitialLogLevel } from "../redux/logStream/logStream.constants"
-import { findIndex, cloneDeep } from "lodash"
+import { cloneDeep } from "lodash"
 import { Level, LogLevelOccurrences } from "../redux/logStream/logStream.types"
 import {
   CustomDataObject,
@@ -14,70 +14,6 @@ import {
 import { isFlowNode, isFlowEdge } from "./flow-chart"
 
 const customData = Object.keys(CustomDataObject)
-
-export const fileToBase64 = (file: File): Promise<string> =>
-  new Promise((resolve, reject) => {
-    const reader = new FileReader()
-    reader.readAsDataURL(file)
-    reader.onload = () => resolve(reader.result)
-    reader.onerror = (error) => reject(error)
-  })
-
-export const mimeTypeFromDataURI = (dataURI: string) =>
-  dataURI.substring(dataURI.indexOf(":") + 1, dataURI.indexOf(";"))
-
-export const formatDebugRequest = async (
-  textQuery: string,
-  files: FileList | null,
-  rows: string[],
-  locations: { [key: string]: string },
-  keys: { [key: string]: string },
-  values: { [key: string]: string }
-) => {
-  const request: any = {
-    data: [],
-    parameters: {},
-  }
-  if (textQuery) {
-    request.data.push({ text: textQuery })
-  }
-  if (files?.length) {
-    for (let file of Array.from(files)) {
-      const uri = await fileToBase64(file)
-      request.data.push({ uri })
-    }
-  }
-  rows.forEach((row) => {
-    const location = locations[row]
-    const key = keys[row]
-    const value = values[row]
-
-    if (!key || !value) return
-
-    let formattedValue = ""
-
-    try {
-      formattedValue = JSON.parse(value)
-    } catch (e) {
-      formattedValue = value
-    }
-
-    if (!location || location === "parameters")
-      request.parameters[key] = formattedValue
-    else if (location === "root") request[key] = formattedValue
-    else if (location === "textQuery" && textQuery)
-      request.data[0][key] = formattedValue
-    else if (files) {
-      let dataIndex = findIndex(
-        Array.from(files),
-        (file) => `file-${file.name}` === location
-      )
-      if (textQuery) dataIndex = 1
-      if (dataIndex >= 0) request.data[dataIndex][key] = formattedValue
-    }
-  })
-  return JSON.stringify(request, null, "\t")
-}
 
 export const parseYAML = (yamlSTR: string) => {
   let yamlStrWithoutTag = /^!/.test(yamlSTR)
